music_platform/pyproject.toml
[tool.poetry]
name = "music-platform"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "music_platform"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.3"
django-model-utils = "^4.2.0"
Pillow = "^9.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
